AutomataKeyGen fetches long instead of ulong

+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Internal/ILGeneratorExtensions.cs
@@ -357,9 +357,9 @@ namespace MessagePack.Internal
             il.Emit(OpCodes.Ret);
         }
 
-        public static void EmitULong(this ILGenerator il, ulong value)
+        public static void EmitULong(this ILGenerator il, long value)
         {
-            il.Emit(OpCodes.Ldc_I8, unchecked((long)value));
+            il.Emit(OpCodes.Ldc_I8, unchecked(value));
         }
 
         public static void EmitThrowNotimplemented(this ILGenerator il)
